requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=13.0
pydantic>=2.9.0
python-dotenv>=1.0.0
tavily-python>=0.5.0
chromadb>=0.5.0
biopython>=1.84
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
